backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow==10.4.0
django-extensions==3.2.3
djoser==2.2.2
PyJWT==2.8.0
requests==2.31.0
openai==1.3.7
stripe==7.6.0
python-multipart==0.0.6
uvicorn==0.24.0
whitenoise==6.6.0
django-filter==23.5
django-redis==5.4.0